fix: address third round of code review findings
- Split fake-gcs-server tests into dedicated file to keep
  terraform_backend_gcs_test.go under 600 lines
- Wrap uploadMetadata errors with ErrPlanfileMetadataFailed sentinel
  so callers can use errors.Is
- Replace fragile index-based relKey slicing with strings.TrimPrefix
  for robustness with trailing-slash prefixes
